Split the app shell's dashboard selectors into their own module
